+++ src/gui/Onyx32.Gui/src/Controls/Button/Button.h
@@ -18,7 +18,7 @@ namespace Onyx32::Gui::Controls
 			virtual LRESULT Process(unsigned int message, WPARAM wParam, LPARAM lParam) override;
 			virtual void SetEvent(ButtonEvents evt, OnButtonEvent&& evtHandler) override;
 
-			static IButton* Create(IWindow* parent, unsigned int controlId, std::wstring_view text, unsigned int width, unsigned int height, unsigned int xPos, unsigned int yPos);
+			static IButton* Create(IWindow* parent, uint64_t controlId, std::wstring_view text, unsigned int width, unsigned int height, unsigned int xPos, unsigned int yPos);
 
 		protected:
 			Button(
@@ -27,7 +27,7 @@ namespace Onyx32::Gui::Controls
 				const unsigned int height = 100,
 				const unsigned int xPos = 100,
 				const unsigned int yPos = 100,
-				const unsigned int controlId = 0);
+				const uint64_t controlId = 0);
 
 			std::wstring _text;
 			std::unordered_map<ButtonEvents, OnButtonEvent> _buttonEventHandlers;
